+++ DiscRental73TestWpf/Infrastructure/HelperModels/EditProductCostModel.cs
@@ -4,7 +4,7 @@
     {
         public int ProductId { get; set; }
         public string DiscTitle { get; set; }
-        public double CurrentCost { get; set; }
-        public double NewCost { get; set; }
+        public decimal CurrentCost { get; set; }
+        public decimal NewCost { get; set; }
     }
 }
